Rename to `gptools`.
