+++ MANIFEST.in
@@ -1,1 +1,1 @@
-include graph_gaussian_process/graph_gaussian_process.stan
+include gptools/gptools.stan
