+++ gptools/examples/poisson_regression/data.stan
