+++ gptools/util.py
