+++ setup.cfg
@@ -10,7 +10,7 @@ exclude_lines =
     except ModuleNotFoundError
     def __repr__
 omit =
-    graph_gaussian_process/__main__.py
+    gptools/__main__.py
 
 [tool:pytest]
 addopts = -v
