+++ setup.py
@@ -2,7 +2,7 @@ from setuptools import find_packages, setup
 
 
 setup(
-    name="graph_gaussian_process",
+    name="gptools",
     packages=find_packages(),
     version="0.1.0",
     install_requires=[
