+++ test_requirements.txt
@@ -39,7 +39,7 @@ charset-normalizer==2.0.7
 cloudpickle==2.1.0
     # via doit
 cmdstanpy==1.0.4
-    # via graph-gaussian-process
+    # via gptools
 colorama==0.4.5
     # via doit-interface
 coverage[toml]==6.0.2
@@ -57,7 +57,7 @@ docutils==0.17.1
 doit==0.36.0
     # via doit-interface
 doit-interface==0.1.6
-    # via graph-gaussian-process
+    # via gptools
 entrypoints==0.4
     # via
     #   jupyter-client
@@ -67,7 +67,7 @@ executing==0.9.1
 fastjsonschema==2.16.1
     # via nbformat
 flake8==4.0.1
-    # via graph-gaussian-process
+    # via gptools
 fonttools==4.34.4
     # via matplotlib
 idna==3.3
@@ -107,7 +107,7 @@ jinja2==3.0.2
 jsonschema==4.9.1
     # via nbformat
 jupyter==1.0.0
-    # via graph-gaussian-process
+    # via gptools
 jupyter-client==7.3.4
     # via
     #   ipykernel
@@ -137,7 +137,7 @@ markupsafe==2.0.1
     #   jinja2
     #   nbconvert
 matplotlib==3.5.3
-    # via graph-gaussian-process
+    # via gptools
 matplotlib-inline==0.1.3
     # via
     #   ipykernel
@@ -164,7 +164,7 @@ nest-asyncio==1.5.5
     #   nbclient
     #   notebook
 networkx==2.8.5
-    # via graph-gaussian-process
+    # via gptools
 notebook==6.4.12
     # via
     #   jupyter
@@ -172,7 +172,7 @@ notebook==6.4.12
 numpy==1.22.1
     # via
     #   cmdstanpy
-    #   graph-gaussian-process
+    #   gptools
     #   matplotlib
     #   pandas
     #   scipy
@@ -235,10 +235,10 @@ pyrsistent==0.18.1
     # via jsonschema
 pytest==6.2.5
     # via
-    #   graph-gaussian-process
+    #   gptools
     #   pytest-cov
 pytest-cov==3.0.0
-    # via graph-gaussian-process
+    # via gptools
 python-dateutil==2.8.2
     # via
     #   jupyter-client
@@ -261,7 +261,7 @@ qtpy==2.2.0
 requests==2.26.0
     # via sphinx
 scipy==1.9.0
-    # via graph-gaussian-process
+    # via gptools
 send2trash==1.8.0
     # via notebook
 six==1.16.0
@@ -274,7 +274,7 @@ snowballstemmer==2.1.0
 soupsieve==2.3.2.post1
     # via beautifulsoup4
 sphinx==4.2.0
-    # via graph-gaussian-process
+    # via gptools
 sphinxcontrib-applehelp==1.0.2
     # via sphinx
 sphinxcontrib-devhelp==1.0.2
@@ -290,7 +290,7 @@ sphinxcontrib-serializinghtml==1.1.5
 stack-data==0.3.0
     # via ipython
 tabulate==0.8.10
-    # via graph-gaussian-process
+    # via gptools
 terminado==0.15.0
     # via notebook
 tinycss2==1.1.1
@@ -300,7 +300,7 @@ toml==0.10.2
 tomli==1.2.1
     # via coverage
 torch==1.12.1
-    # via graph-gaussian-process
+    # via gptools
 tornado==6.2
     # via
     #   ipykernel
@@ -310,7 +310,7 @@ tornado==6.2
 tqdm==4.64.0
     # via
     #   cmdstanpy
-    #   graph-gaussian-process
+    #   gptools
 traitlets==5.3.0
     # via
     #   ipykernel
