+++ tests/test_examples.py
@@ -7,7 +7,7 @@ from unittest import mock
 
 
 @pytest.mark.filterwarnings("ignore::DeprecationWarning")
-@pytest.mark.parametrize("filename", glob.glob("graph_gaussian_process/examples/*/*.ipynb"))
+@pytest.mark.parametrize("filename", glob.glob("gptools/examples/*/*.ipynb"))
 def test_example(filename: str) -> None:
     with open(filename) as fp:
         notebook = nbformat.read(fp, as_version=4)
