+++ tests/test_stan.py
@@ -1,7 +1,7 @@
 import pathlib
-from graph_gaussian_process import stan
+from gptools import stan
 
 
 def test_include() -> None:
-    include = pathlib.Path(stan.get_include()) / "graph_gaussian_process.stan"
+    include = pathlib.Path(stan.get_include()) / "gptools.stan"
     assert include.is_file()
